Player plumbed
But there is a super weird play/pause bug that comes into play if you
have selected more than one song and then play/pause. It toggles the
state on all the items. In the works to fix it but wanted to get access
to the basic player functionality asap.